Bug is related to future::is_ready() being unstable
It looks like the on_response_ready future is getting called, but
both the timer future and the task_completion futures are returning
false when I call future::is_ready(). The boost::future documentation
says about the function: "The result of this function is not stable
and the future could become not ready even if the function returned
true or vice-versa." Which sucks and makes this function kind of
useless. This would also explain why it happens on mac but not Docker,
different OS's leading to different undefined behavior.